block tests rename
